+++ 5_More/EFCore/InhertianceMappingWithConversion/Usings.cs
@@ -1,0 +1,6 @@
+﻿global using Microsoft.EntityFrameworkCore;
+global using Microsoft.Extensions.Configuration;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Hosting;
+
+global using System.Diagnostics.CodeAnalysis;
